shift things around
